+++ scalingomock/addonsservice_mock.go
@@ -5,6 +5,7 @@
 package scalingomock
 
 import (
+	context "context"
 	reflect "reflect"
 
 	scalingo "github.com/Scalingo/go-scalingo/v4"
@@ -35,105 +36,105 @@ func (m *MockAddonsService) EXPECT() *MockAddonsServiceMockRecorder {
 }
 
 // AddonDestroy mocks base method.
-func (m *MockAddonsService) AddonDestroy(arg0, arg1 string) error {
+func (m *MockAddonsService) AddonDestroy(arg0 context.Context, arg1, arg2 string) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AddonDestroy", arg0, arg1)
+	ret := m.ctrl.Call(m, "AddonDestroy", arg0, arg1, arg2)
 	ret0, _ := ret[0].(error)
 	return ret0
 }
 
 // AddonDestroy indicates an expected call of AddonDestroy.
-func (mr *MockAddonsServiceMockRecorder) AddonDestroy(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockAddonsServiceMockRecorder) AddonDestroy(arg0, arg1, arg2 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonDestroy", reflect.TypeOf((*MockAddonsService)(nil).AddonDestroy), arg0, arg1)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonDestroy", reflect.TypeOf((*MockAddonsService)(nil).AddonDestroy), arg0, arg1, arg2)
 }
 
 // AddonLogsArchives mocks base method.
-func (m *MockAddonsService) AddonLogsArchives(arg0, arg1 string, arg2 int) (*scalingo.LogsArchivesResponse, error) {
+func (m *MockAddonsService) AddonLogsArchives(arg0 context.Context, arg1, arg2 string, arg3 int) (*scalingo.LogsArchivesResponse, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AddonLogsArchives", arg0, arg1, arg2)
+	ret := m.ctrl.Call(m, "AddonLogsArchives", arg0, arg1, arg2, arg3)
 	ret0, _ := ret[0].(*scalingo.LogsArchivesResponse)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // AddonLogsArchives indicates an expected call of AddonLogsArchives.
-func (mr *MockAddonsServiceMockRecorder) AddonLogsArchives(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockAddonsServiceMockRecorder) AddonLogsArchives(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonLogsArchives", reflect.TypeOf((*MockAddonsService)(nil).AddonLogsArchives), arg0, arg1, arg2)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonLogsArchives", reflect.TypeOf((*MockAddonsService)(nil).AddonLogsArchives), arg0, arg1, arg2, arg3)
 }
 
 // AddonLogsURL mocks base method.
-func (m *MockAddonsService) AddonLogsURL(arg0, arg1 string) (string, error) {
+func (m *MockAddonsService) AddonLogsURL(arg0 context.Context, arg1, arg2 string) (string, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AddonLogsURL", arg0, arg1)
+	ret := m.ctrl.Call(m, "AddonLogsURL", arg0, arg1, arg2)
 	ret0, _ := ret[0].(string)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // AddonLogsURL indicates an expected call of AddonLogsURL.
-func (mr *MockAddonsServiceMockRecorder) AddonLogsURL(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockAddonsServiceMockRecorder) AddonLogsURL(arg0, arg1, arg2 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonLogsURL", reflect.TypeOf((*MockAddonsService)(nil).AddonLogsURL), arg0, arg1)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonLogsURL", reflect.TypeOf((*MockAddonsService)(nil).AddonLogsURL), arg0, arg1, arg2)
 }
 
 // AddonProvision mocks base method.
-func (m *MockAddonsService) AddonProvision(arg0 string, arg1 scalingo.AddonProvisionParams) (scalingo.AddonRes, error) {
+func (m *MockAddonsService) AddonProvision(arg0 context.Context, arg1 string, arg2 scalingo.AddonProvisionParams) (scalingo.AddonRes, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AddonProvision", arg0, arg1)
+	ret := m.ctrl.Call(m, "AddonProvision", arg0, arg1, arg2)
 	ret0, _ := ret[0].(scalingo.AddonRes)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // AddonProvision indicates an expected call of AddonProvision.
-func (mr *MockAddonsServiceMockRecorder) AddonProvision(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockAddonsServiceMockRecorder) AddonProvision(arg0, arg1, arg2 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonProvision", reflect.TypeOf((*MockAddonsService)(nil).AddonProvision), arg0, arg1)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonProvision", reflect.TypeOf((*MockAddonsService)(nil).AddonProvision), arg0, arg1, arg2)
 }
 
 // AddonToken mocks base method.
-func (m *MockAddonsService) AddonToken(arg0, arg1 string) (string, error) {
+func (m *MockAddonsService) AddonToken(arg0 context.Context, arg1, arg2 string) (string, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AddonToken", arg0, arg1)
+	ret := m.ctrl.Call(m, "AddonToken", arg0, arg1, arg2)
 	ret0, _ := ret[0].(string)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // AddonToken indicates an expected call of AddonToken.
-func (mr *MockAddonsServiceMockRecorder) AddonToken(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockAddonsServiceMockRecorder) AddonToken(arg0, arg1, arg2 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonToken", reflect.TypeOf((*MockAddonsService)(nil).AddonToken), arg0, arg1)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonToken", reflect.TypeOf((*MockAddonsService)(nil).AddonToken), arg0, arg1, arg2)
 }
 
 // AddonUpgrade mocks base method.
-func (m *MockAddonsService) AddonUpgrade(arg0, arg1 string, arg2 scalingo.AddonUpgradeParams) (scalingo.AddonRes, error) {
+func (m *MockAddonsService) AddonUpgrade(arg0 context.Context, arg1, arg2 string, arg3 scalingo.AddonUpgradeParams) (scalingo.AddonRes, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AddonUpgrade", arg0, arg1, arg2)
+	ret := m.ctrl.Call(m, "AddonUpgrade", arg0, arg1, arg2, arg3)
 	ret0, _ := ret[0].(scalingo.AddonRes)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // AddonUpgrade indicates an expected call of AddonUpgrade.
-func (mr *MockAddonsServiceMockRecorder) AddonUpgrade(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockAddonsServiceMockRecorder) AddonUpgrade(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonUpgrade", reflect.TypeOf((*MockAddonsService)(nil).AddonUpgrade), arg0, arg1, arg2)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonUpgrade", reflect.TypeOf((*MockAddonsService)(nil).AddonUpgrade), arg0, arg1, arg2, arg3)
 }
 
 // AddonsList mocks base method.
-func (m *MockAddonsService) AddonsList(arg0 string) ([]*scalingo.Addon, error) {
+func (m *MockAddonsService) AddonsList(arg0 context.Context, arg1 string) ([]*scalingo.Addon, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AddonsList", arg0)
+	ret := m.ctrl.Call(m, "AddonsList", arg0, arg1)
 	ret0, _ := ret[0].([]*scalingo.Addon)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // AddonsList indicates an expected call of AddonsList.
-func (mr *MockAddonsServiceMockRecorder) AddonsList(arg0 interface{}) *gomock.Call {
+func (mr *MockAddonsServiceMockRecorder) AddonsList(arg0, arg1 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonsList", reflect.TypeOf((*MockAddonsService)(nil).AddonsList), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddonsList", reflect.TypeOf((*MockAddonsService)(nil).AddonsList), arg0, arg1)
 }